send_review_notification/requirements.txt
python-dateutil==2.8.1
requests==2.26.0
